+++ src/components/BarChart/stories/BarChart.stories.tsx
@@ -148,3 +148,31 @@ LastBarTreatment.args = {
   yAxisOptions: {labelFormatter: formatYAxisLabel},
   renderTooltipContent,
 };
+
+export const MinimalLabels = Template.bind({});
+MinimalLabels.args = {
+  data: [
+    {rawValue: 1324.19, label: '1 day some really long long text'},
+    {rawValue: 1022.79, label: '2020-01-02T12:00:00Z'},
+    {rawValue: 713.29, label: '2020-01-03T12:00:00Z'},
+    {rawValue: 413.29, label: '2020-01-04T12:00:00Z'},
+    {
+      rawValue: 100.79,
+      label:
+        'Middle day also has very long text / Middle day also has very long text',
+    },
+    {rawValue: 350.6, label: '2020-01-06T12:00:00Z'},
+    {rawValue: 277.69, label: '2020-01-07T12:00:00Z'},
+    {
+      rawValue: 50.6,
+      label: 'Last day has especially longgggggggggggg textttttttttt',
+    },
+  ],
+  barOptions: {
+    color: barGradient,
+    hasRoundedCorners: true,
+  },
+  xAxisOptions: {useMinimalLabels: true},
+  yAxisOptions: {labelFormatter: formatYAxisLabel},
+  renderTooltipContent,
+};